pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tiktok-collector"
version = "0.1.0"
description = "Public TikTok metadata collector"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "playwright>=1.52.0",
  "pandas>=2.2.2",
  "PyYAML>=6.0.2",
  "python-dotenv>=1.0.1",
  "yt-dlp>=2026.3.13",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
tiktok-collect = "tiktok_collector.cli:main"
